CHROMIUM: cpufreq: interactive: calculate load before freq change
The update to cpu load for cpufreq rate changes was happening after the
rate
change instead of before. This switches it to update the cpu load
before the
cpufreq rate change so the old frequency is used for
calculating the speed
adjusted cpu load. The old frequency was used for
that active time, so it should
be used for calculating the speed
adjusted cpu load.

BUG=chrome-os-partner:37673
TEST=power_LoadTest on
Jerry
check we spend most of our time at the lower cpu
frequencies

Change-Id:
I2fff785561ad8aebd354ddd305e91d96c799a617
Signed-off-by: Derek Basehore
<dbasehore@chromium.org>
Reviewed-on:
https://chromium-review.googlesource.com/262925
Reviewed-by: Douglas
Anderson <dianders@chromium.org>